move alert code to separate file
