+++ macadamia/Alerts.swift
@@ -1,0 +1,61 @@
+//
+//  Alerts.swift
+//  macadamia
+//
+//  Created by zm on 12.11.24.
+//
+
+import Foundation
+import SwiftUI
+
+struct AlertDetail {
+    let title: String
+    let alertDescription: String?
+    let primaryButtonText: String?
+    let affirmText: String?
+    let onAffirm: (() -> Void)?
+
+    init(title: String,
+         description: String? = nil,
+         primaryButtonText: String? = nil,
+         affirmText: String? = nil,
+         onAffirm: (() -> Void)? = nil)
+    {
+        self.title = title
+        alertDescription = description
+        self.primaryButtonText = primaryButtonText
+        self.affirmText = affirmText
+        self.onAffirm = onAffirm
+    }
+}
+
+struct AlertViewModifier: ViewModifier {
+    @Binding var isPresented: Bool
+    var currentAlert: AlertDetail?
+
+    func body(content: Content) -> some View {
+        content
+            .alert(currentAlert?.title ?? "Error", isPresented: $isPresented) {
+                Button(role: .cancel) {
+                    // This button could potentially reset or handle cancel logic
+                } label: {
+                    Text(currentAlert?.primaryButtonText ?? "OK")
+                }
+                if let affirmText = currentAlert?.affirmText, let onAffirm = currentAlert?.onAffirm {
+                    Button(role: .destructive) {
+                        onAffirm()
+                    } label: {
+                        Text(affirmText)
+                    }
+                }
+            } message: {
+                Text(currentAlert?.alertDescription ?? "")
+            }
+    }
+}
+
+extension View {
+    func alertView(isPresented: Binding<Bool>, currentAlert: AlertDetail?) -> some View {
+        modifier(AlertViewModifier(isPresented: isPresented, currentAlert: currentAlert))
+    }
+}
